split vanilla.py into subfiles for clarity
